Breakup old utils test to new structure
